Move build_executable_project helper to test-support crate
