+++ extensions/scarb-verify/tests/build.rs
@@ -1,29 +1,6 @@
-use assert_fs::TempDir;
 use indoc::indoc;
 use scarb_test_support::command::Scarb;
-use scarb_test_support::project_builder::ProjectBuilder;
-
-fn build_executable_project() -> TempDir {
-    let t = TempDir::new().unwrap();
-    ProjectBuilder::start()
-        .name("hello")
-        .version("0.1.0")
-        .dep_cairo_execute()
-        .manifest_extra(indoc! {r#"
-                [executable]
-
-                [cairo]
-                enable-gas = false
-            "#})
-        .lib_cairo(indoc! {r#"
-            #[executable]
-            fn main() -> felt252 {
-                42
-            }
-        "#})
-        .build(&t);
-    t
-}
+use scarb_test_support::fixtures::build_executable_project;
 
 #[test]
 fn verify_fails_when_execution_output_not_found() {
